A proof of the splitting lemma for Stacks
The goal of this patch is to establish the well-known splitting lemma,
currently [Lemma 010G](https://stacks.math.columbia.edu/tag/010G) (`lemma-ses-split`)
with proof omitted in [Section 00ZX](https://stacks.math.columbia.edu/tag/00ZX) (`section-abelian-categories`).

However, the theorem can be stated in a manner which is equivalent for Abelian categories
and true in all preadditive categories; cf `lemma-split-morphism-splitting` below.

Our approach to proving the theorem is to first show (in the full generality of the preadditive setting)
the well-known splitting result for idempotents
(cf. [Lemma 09SH](https://stacks.math.columbia.edu/tag/09SH), `lemma-karoubian`).
This involves the creation of 5 labels:
- 1 preliminary remark in `section-additive-categories` concerning idempotents
- 2 lemmate in `section-additive-categories` concerning idempotents
- 2 lemmate in `section-additive-categories` concerning section/retraction pairs.
As an incidental bonus we are able to complete and simplify the proof of the aforementioned
[Lemma 09SH](https://stacks.math.columbia.edu/tag/09SH) (`lemma-karoubian`).

There is room to consider merging created lemmata; cf. the below lists:

**Labels created:**
- `lemma-idempotent-symmetry`: The complement of an idempotent is idempotent, and the two annihilate one another.
- `lemma-idempotent-kernel-cokernel`: An idempotent has a kernel iff it has a cokernel, and likewise for its complement.
- `lemma-idempotent-splitting`: Idempotents beget direct product decompositions of their (co)domains.
- `lemma-split-morphism-kernel-cokernel`: The section/retraction of a split pair of morphisms is a kernel/cokernel respectively.
- `lemma-split-morphism-splitting`: The section/retraction of a split pair of morphisms beget a splitting of the codomain/domain respectively.

**Labels modified:**
- `lemma-karoubian`: Proof is completed ("omit" eliminated) and simplified with the claim an immediate corollary of the above and [Lemma 09QG](https://stacks.math.columbia.edu/tag/09QG) (`lemma-additive-cat-biproduct-kernel`)
- `lemma-ses-split`: Proof is added; the claim is trivialized via the created lemmata.